Backend/requirements.txt

llama-index==0.10.43
langchain==0.1.16
pydantic==2.7.1
fastapi==0.111.0
uvicorn==0.29.0
faiss-cpu
openai==0.28.1
python-dotenv==1.0.1
